fixed comments

+++ db/migrate/20241029210702_drop_sessions_table.rb
@@ -1,6 +1,6 @@
 class DropSessionsTable < ActiveRecord::Migration[7.2]
   def up
-    drop_table :sessions
+    drop_table :sessions if table_exists?(:sessions)
   end
 
   def down
